+++ tests/modules/admin/ApiAdminStorageTest.php
@@ -11,8 +11,7 @@ class ApiAdminStorageTest extends HeadlessTestCase
     public function testImageUpload()
     {
         $image = __DIR__ . '/../../data/image.png';
-
-        $upload = ApiAdminStorage::imageUpload($image, 'image/jpg', 'image.jpg')->response($this->createDummyClient([
+        $client = $this->createDummyClient([
             'image' => [
                 'id' => 1,
                 'filter_id' => 0,
@@ -23,22 +22,23 @@ class ApiAdminStorageTest extends HeadlessTestCase
             'mediumThumbnail' => [
 
             ]
-        ]));
+        ]);
+        $upload = ApiAdminStorage::imageUpload($client, $image, 'image/jpg', 'image.jpg');
 
-        $this->assertNotNull($upload->getContent());
+        $this->assertInstanceOf('luya\headless\modules\admin\ApiStorageImage', $upload);
     }
 
     public function testFileUpload()
     {
         $image = __DIR__ . '/../../data/image.png';
-
-        $upload = ApiAdminStorage::fileUpload($image, 'image/jpg', 'image.jpg')->response($this->createDummyClient([
+        $client = $this->createDummyClient([
             'file' => [
                 'id' => 1,
                 'source' => 'path/to/image/on/server/png',
             ],
-        ]));
+        ]);
+        $upload = ApiAdminStorage::fileUpload($client, $image, 'image/jpg', 'image.jpg');
 
-        $this->assertNotNull($upload->getContent());
+        $this->assertInstanceOf('luya\headless\modules\admin\ApiStorageFile', $upload);
     }
 }